requirements.txt
pygame-ce>=2.5.4
pillow>=12.0.0
pygbag>=0.9.2